sales_engagement_action/requirements.txt
# Core dependencies
anthropic>=0.39.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0

# From Scripts 1 & 2 (for direct integration)
requests>=2.31.0
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
hashlib-additional>=1.0.0
